Bean Validation

+++ src/main/java/com/devsuperior/dscommerce/services/ProductService.java
@@ -3,7 +3,11 @@ package com.devsuperior.dscommerce.services;
 import com.devsuperior.dscommerce.dto.ProductDTO;
 import com.devsuperior.dscommerce.entities.Product;
 import com.devsuperior.dscommerce.repositories.ProductRepository;
+import com.devsuperior.dscommerce.services.exceptions.DatabaseException;
+import com.devsuperior.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -19,7 +23,7 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
-        Product product = repository.findById(id).get();
+        Product product = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado. "));
         return new ProductDTO(product);
     }
 
@@ -42,18 +46,42 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto) {
-        Product entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
 
-        entity = repository.save(entity);
+        try {
+            Product entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
 
-        return new ProductDTO(entity);
+            entity = repository.save(entity);
+
+            return new ProductDTO(entity);
+        }
+        catch (EntityNotFoundException e) {
+
+            throw new ResourceNotFoundException("Recurso não encontrado.");
+
+        }
+
 
     }
 
     @Transactional
     public void delete(Long id) {
-     repository.deleteById(id);
+
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Id nao existe no banco de dados");
+        }
+
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+
+        }
+
+
+
+
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
